chore: moved vault to OSS
