+++ README.md
@@ -73,7 +73,9 @@ Flipt v2 is the first truly Git-native feature management platform that treats y
 ### 🔒 **Enterprise Security & Control**
 
 - **Self-hosted**: Keep sensitive flag data within your infrastructure
-- **Merge proposals**: Code review workflow for flag changes (commercial feature)
+- **Secrets management**: Secure storage and retrieval of sensitive configuration data (OSS)
+- **GPG commit signing**: Cryptographically sign all flag changes for enhanced security (Pro feature)
+- **Merge proposals**: Code review workflow for flag changes (Pro feature)
 - **Audit trails**: Complete history of who changed what and when
 - **OIDC/JWT/OAuth**: Enterprise authentication methods supported
 
@@ -144,13 +146,31 @@ Flipt UI will be available at [http://127.0.0.1:8080/](http://127.0.0.1:8080).
 ### Configuration Example
 
 ```yaml
-# config.yml - Git-native setup
+# config.yml - Git-native setup with secrets management
+secrets:
+  providers:
+    vault:
+      enabled: true
+      address: "https://vault.example.com"
+      auth_method: "token"
+      token: "hvs.your_token"
+      mount: "secret"
+
 storage:
   type: git
   git:
     repository: "https://github.com/your-org/feature-flags.git"
     ref: "main"
     poll_interval: "30s"
+    signature:
+      enabled: true
+      type: "gpg"
+      key_ref:
+        provider: "vault"
+        path: "flipt/signing-key"
+        key: "private_key"
+      name: "Flipt Bot"
+      email: "bot@example.com"
 
 environments:
   default:
@@ -167,6 +187,7 @@ For more setup options, see our [configuration documentation](https://docs.flipt
 ## Core Values
 
 - 🔒 **Security** - HTTPS, OIDC, JWT, OAuth, K8s Service Token, and API Token authentication methods supported out of the box
+- 🗝️ **Secrets Management** - Secure storage and retrieval of sensitive data with HashiCorp Vault and file-based providers
 - 🎛️ **Control** - Your data stays in your Git repositories within your infrastructure  
 - 🚀 **Speed** - Co-located with your services, no external API calls required
 - ✅ **Simplicity** - Single binary with no external dependencies by default
@@ -182,12 +203,20 @@ For more setup options, see our [configuration documentation](https://docs.flipt
 - Store flags directly in Git repositories alongside your code
 - Full version control with Git history, blame, and diff support  
 - Integrates with your SCM (GitHub, GitLab, Gitea, etc.)
+- GPG commit signing for cryptographic verification of changes
 
 ### Multi-Environment Management  
 
 - Environment per Git branch, directory, or repository
 - Complete environment isolation with independent configurations
 - Seamless environment promotion workflows
+
+### Secrets Management & Security
+
+- **Multi-provider secrets management**: File-based and HashiCorp Vault providers available in OSS
+- **GPG commit signing**: Cryptographically sign all flag changes with keys from secret providers (Pro feature)
+- **Secure key storage**: Private keys and sensitive data stored securely in Vault or local files
+- **Multiple auth methods**: Token, Kubernetes, and AppRole authentication for Vault
 
 ### Advanced Flag Management
 
@@ -203,9 +232,11 @@ For more setup options, see our [configuration documentation](https://docs.flipt
 
 ### Enterprise Features
 
-- Merge proposals for flag changes (commercial feature)
-- Authentication via OIDC, JWT, OAuth, and more
-- OpenTelemetry and Prometheus integration 🔋
+- **Secrets Management**: Secure storage with HashiCorp Vault and file-based providers (OSS)
+- **GPG Commit Signing**: Cryptographically sign all flag changes for enhanced security (Pro feature)
+- **Merge proposals**: Code review workflow for flag changes (Pro feature)
+- **Authentication**: OIDC, JWT, OAuth, and more authentication methods
+- **Observability**: OpenTelemetry and Prometheus integration 🔋
 
 <br clear="both"/>
 
@@ -261,6 +292,7 @@ Ready to unlock the full potential of Git-native feature management? Flipt v2 Pr
 ### What's Included in Pro
 
 - **🔀 Merge Proposals** - Code review workflow for feature flag changes, just like GitHub PRs
+- **✍️ GPG Commit Signing** - Cryptographically sign all changes using keys from secret providers for enhanced security and auditability
 - **🏢 Premium Support** - Shared Slack channel with same-day response times
 - **⚡ Priority Development** - Your bug reports and feature requests get prioritized
 - **🔧 Enterprise Auth** - Advanced authentication providers (coming soon)
