+++ internal/cmd/grpc.go
@@ -23,8 +23,8 @@ import (
 	"go.flipt.io/flipt/internal/secrets"
 
 	// Import providers to trigger their init functions
-	_ "go.flipt.io/flipt/internal/coss/secrets/vault"
 	_ "go.flipt.io/flipt/internal/secrets/file"
+	_ "go.flipt.io/flipt/internal/secrets/vault"
 	serverfliptv1 "go.flipt.io/flipt/internal/server"
 	analytics "go.flipt.io/flipt/internal/server/analytics"
 	"go.flipt.io/flipt/internal/server/analytics/clickhouse"
